which-bollywood-celebrity-are-you/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib-metadata==6.6.0
jax==0.4.12
keras==2.12.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-vggface==0.6
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mtcnn==0.1.0
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.12.1
zipp==3.15.0
